metric: fix calculation of average bitrate

+++ src/core.h
@@ -250,7 +250,7 @@ struct metric {
 void     metric_init(struct metric *metric);
 void     metric_reset(struct metric *metric);
 void     metric_add_packet(struct metric *metric, size_t packetsize);
-uint32_t metric_avg_bitrate(const struct metric *metric);
+double   metric_avg_bitrate(const struct metric *metric);
 
 
 /*
